big changes

+++ src/gpki.cpp
@@ -1,15 +1,19 @@
 #include "gpki.hpp"
 #include "logger/signals.hpp"
 #include "parse/parser.hpp"
-#include "utils/utils.hpp"
-
-
-#include "printing/printing.cpp"
-#include <cstdlib>
-#include <filesystem>
-#include <iostream>
+#include "db/entities.hpp"
 
 using namespace gpkih;
+
+constexpr static size_t __dynamic_memory_size = 1024 * 4096; // 4MB
+
+/* Error */
+std::string last_gpkih_error = "";
+void printlasterror() {
+    if (!last_gpkih_error.empty()) {
+        fmt::print(fg(RED), last_gpkih_error + "\n");
+    }
+}
 
 /* Directory names */
 static constexpr const char* DB_DIRNAME  = "db";
@@ -17,41 +21,39 @@ static constexpr const char* CFG_DIRNAME = "config";
 static constexpr const char* LOG_DIRNAME = "logs";
 
 /* Configuration names */
-str vpn_conf_filename = "openvpn.conf";
-str pki_conf_filename = "pki.conf";
+std::string vpn_conf_filename = "openvpn.conf";
+std::string pki_conf_filename = "pki.conf";
 
 /* Current path */
-static str CURRENT_PATH;
+static std::string CURRENT_PATH;
 
-static str GPKIH_BASEDIR; // initialized by __set_globals()
-str CONF_DIRPATH;         // proper value set by __set_globals()
-static str DB_DIRPATH;    // initialized by __set_globals()
+std::string GPKIH_BASEDIR = "";   // proper value set by __set_platform_dependant_variables()
+std::string CONF_DIRPATH = "";    // proper value set by __set_variables()
+std::string DB_DIRPATH = "";      // proper value set by __set_variables()
 
-static str DB_PROFILES_CSV;
+static std::string CONF_GPKIH;    // initialized by __set_variables()s
 
-static str CONF_GPKIH;    // initialized by __set_globals()
+static std::string gpkih_conf_filename = "gpkih.conf";
+static std::string LOG_DIRPATH;
 
-static str gpkih_conf_filename = "gpkih.conf";
-static str LOG_DIRPATH;
+char *serial_path = NULL;
+uint64_t serial_path_len = 0;
 
-static str CSV_DELIMITER_s = ",";
-static char CSV_DELIMITER_c = ',';
+static int __check_gpkih_install_dir(std::string &path) {
+    PDEBUG(1, "__check_gpkih_install_dir()");
 
-static int check_gpkih_install_dir(str &path) {
     if (!fs::exists(path))
     {
-        str ans;
-        PROMPT("About to create gpkih source dir , proceed?", "[y/n]");
-        getline(std::cin, ans);
-        for (char& c : ans) { c = tolower(c); }
-        if (ans != "y" && ans != "yes") {
-            PINFO("User refused, exiting ...\n");
-            return GPKIH_FAIL;
+        if(Config::get("behaviour","autoasnwer") == "no"){
+            auto ans = PROMPT("About to create gpkih source dir , proceed?", "[y/n]", true);
+            if (ans != "y" && ans != "yes") {
+                PINFO("User refused, exiting ...\n");
+                return GPKIH_FAIL;
+            }
         }
 
-        PINFO("Creating gpkih source dir '{}'\n", path);
         if (!fs::create_directory(path)) {
-            seterror("Couldn't create directory '{}'\n", path);
+            seterror("Couldn't create gpkih's root directory '{}'\n", path);
             return F_NOEXIST;
         };
 
@@ -60,7 +62,13 @@ static int check_gpkih_install_dir(str &path) {
         //      [WINDOWS] $env:localappadata\gpkih 
         //      [LINUX] ~/.config/gpkih 
         // doesn't exist
-        str configdir = CURRENT_PATH + SLASH + "config";
+        std::string configdir = CURRENT_PATH + SLASH + "config";
+
+        // Check if configdir exists, else we are not in the proper location, exit.
+        if (fs::exists(configdir) == false) {
+            seterror("couldn't find config dir in current directory '{}', exiting...\n");
+            return GPKIH_FAIL;
+        }
 
         // Copy config directory
         fs::copy(configdir, CONF_DIRPATH, fs::copy_options::recursive);
@@ -82,6 +90,8 @@ static int check_gpkih_install_dir(str &path) {
 
 static int __set_platform_dependant_variables()
 {
+    PDEBUG(1,"__set_platform_dependant_variables()");
+
     #ifdef _WIN32
     GPKIH_BASEDIR = utils::env::get_environment_variable("LOCALAPPDATA");
     #else
@@ -103,7 +113,8 @@ static int __set_platform_dependant_variables()
     return GPKIH_OK;
 }
 
-static int set_variables() {
+static int __set_variables() {
+    PDEBUG(1,"__set_variables()");
 
     if (__set_platform_dependant_variables() != GPKIH_OK) {
         return GPKIH_FAIL;
@@ -114,39 +125,55 @@ static int set_variables() {
     CONF_DIRPATH    = fmt::format("{}{}{}", GPKIH_BASEDIR, CFG_DIRNAME, SLASH);
     LOG_DIRPATH     = fmt::format("{}{}{}", GPKIH_BASEDIR, LOG_DIRNAME, SLASH);
     CONF_GPKIH      = fmt::format("{}{}", CONF_DIRPATH, gpkih_conf_filename);
-    DB_PROFILES_CSV = fmt::format("{}{}", CONF_DIRPATH, "profiles.csv");
 
-    // db::profiles::dbpath
-    db::profiles::dbpath = fmt::format("{}profiles.csv", DB_DIRPATH);
+    // file used by db::profiles namespace to store/load profiles
+    db::profiles::dbpath = fmt::format("{}profiles.data", DB_DIRPATH);
+
+    // simple id tracking file (increased by 1)
+    db::profiles::idfile = fmt::format("{}id.data", DB_DIRPATH);
+
+    EntityManager::db = DB_DIRPATH;
     
     return GPKIH_OK;
 }
 
 // PROGRAM ENTRY POINT
 int main(int argc, const char **args) {
+  std::vector <std::string> opts(args+1,args+argc);
 
-  for(int i = 0; i < argc; ++i){
-    if(!strcmp(args[i], "--debug")){
-        ENABLE_DEBUG_MESSAGES = true;
-        args[i] = NULL;
-        --argc;
+  for(int i = 0; i < opts.size(); ++i){
+    if(opts[i] == "--debug"){
+        if(i+1 == opts.size()){
+            PERROR("debug level must be given");
+            return GPKIH_FAIL;
+        }
+
+        int debug_level = std::stoi(opts[i+1]);
+        if(debug_level > 0 && debug_level <= 3){
+            ENABLE_DEBUG_MESSAGES = true;
+            DEBUG_LEVEL = debug_level;
+        }else{
+            PERROR("debug level must be 0-3 (both included)");
+            return GPKIH_FAIL;
+        }
+
+        opts.erase(opts.begin()+i,opts.begin()+i+2);
     }
   }
 
-  PDEBUG("setting global variables");
-  if(set_variables() != GPKIH_OK){
+  Buffer a = Buffer(__dynamic_memory_size);
+  __buff = &a;
+
+  if(__set_variables() != GPKIH_OK){
     printlasterror();
     return GPKIH_FAIL;
   };
 
-  
-  
-  PDEBUG("checking base dir");
-  if (check_gpkih_install_dir(GPKIH_BASEDIR) != GPKIH_OK) {
+
+  if (__check_gpkih_install_dir(GPKIH_BASEDIR) != GPKIH_OK) {
     printlasterror();
     return GPKIH_FAIL;
   }
-
 
   // Register signal handlers   
   Signals::register_signals();
@@ -156,31 +183,29 @@ int main(int argc, const char **args) {
   // Launch task to load `gpkih.conf` configuration
   auto load_gpkih_config = std::async(std::launch::async, Config::load, CONF_GPKIH);
 
-  // Map profiles' csv to db::profiles::existing_profiles{}
-  size_t profile_count = db::profiles::initialize(DB_DIRPATH + "profiles.csv");
-  if (profile_count < 0) {
+  // Initialize profiles' db
+  size_t profile_count = 0;
+  if (db::profiles::initialize(profile_count) == GPKIH_FAIL) {
     printlasterror();
+    load_gpkih_config.wait();
     return -1;
   };
 
-  // wait for Config::load to finish
+  // Wait for Config::load to finish
   if (load_gpkih_config.get() != GPKIH_OK) {
     printlasterror();
     return -1;
   }
-  
-  // Initialize the program logger
-  if (Logger::initialize(std::move((fs::path(GPKIH_BASEDIR)/"logs").string())) == false) {
-    printlasterror();
-    return GPKIH_FAIL;
-  };
 
-  auto logger = Logger::get();
+  /* Initialize program logger */
+  Logger::set_basedir(std::move(fs::path(GPKIH_BASEDIR)/"logs").string());
+  Logger pl = Logger("gpkih");
+  gpkih_logger = &pl;
 
-  PDEBUG("Loaded [{}] profiles", profile_count, DB_DIRPATH, SLASH);
+  PDEBUG(1,"Loaded [{}] profiles", profile_count);
 
   // Parse options
-  if (parsers::parse(argc - 1, args + 1) != GPKIH_OK) {
+  if (parsers::parse(opts) != GPKIH_OK) {
     printlasterror();
     return GPKIH_FAIL;
   }
